updated smoke test

+++ benchmark.sh
@@ -34,7 +34,9 @@ then
     python3 experiments-dts-cav.py --paynt-dir /opt/paynt --models-dir ./benchmarks --experiment-name paynt-smoke-test --workers $thread_count --depth-max 1 --generate-csv --smoke-test --restart
 
     echo "generating OMDT log files"
-    rm logs/omdt-smoke-test/results.csv
+    if [ -f logs/omdt-smoke-test/results.csv ]; then
+        rm logs/omdt-smoke-test/results.csv
+    fi
     cd /opt/OMDT
     python3 experiments-dts-cav-omdt.py --omdt-dir ./ --models-dir ./models --experiment-name omdt-smoke-test --workers $thread_count --depth-max 1 --restart
     cd -
@@ -96,7 +98,9 @@ then
     python3 experiments-dts-cav.py --paynt-dir /opt/paynt --models-dir ./benchmarks --experiment-name paynt-cav-final --generate-csv --workers $thread_count --restart
 
     echo "generating OMDT log files"
-    rm logs/omdt-cav-final/results.csv
+    if [ -f logs/omdt-cav-final/results.csv ]; then
+        rm logs/omdt-cav-final/results.csv
+    fi
     cd /opt/OMDT
     python3 experiments-dts-cav-omdt.py --omdt-dir ./ --models-dir ./models --experiment-name omdt-cav-final --workers $thread_count --restart
     cd -
